+++ src/models/sensorWidgetModel.js
