+++ src/routes/sensorWidgetRoutes.js
