+++ src/routes/weatherRoutes.js
