+++ tests/sensorDataRange.test.js
@@ -6,10 +6,10 @@ const FakeModel = {
   async find(q){ this.query=q; return [{_id:'1'}]; }
 };
 
-const modelPath = path.join(__dirname,'..','models','sensorDataModel.js');
+const modelPath = path.join(__dirname,'..','src/models','sensorDataModel.js');
 require.cache[modelPath] = { exports: FakeModel };
 
-const { getSensorDataRange } = require('../controllers/sensorDataController');
+const { getSensorDataRange } = require('../src/controllers/sensorDataController');
 
 function resMock(){ return { statusCode:0,data:null,status(c){this.statusCode=c;return this;},json(d){this.data=d;} }; }
